+++ packages/graph-explorer/src/modules/NodeExpand/NodeExpandContent.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import type { Vertex } from "../../@types/entities";
-import { ModuleContainerFooter } from "../../components";
+import { ModuleContainerFooter, PanelError } from "../../components";
 import Button from "../../components/Button";
 import ExpandGraphIcon from "../../components/icons/ExpandGraphIcon";
 import GraphIcon from "../../components/icons/GraphIcon";
@@ -15,37 +15,66 @@ import defaultStyles from "./NodeExpandContent.styles";
 import NodeExpandFilters, { NodeExpandFilter } from "./NodeExpandFilters";
 import VertexHeader from "../common/VertexHeader";
 import { ExpandNodeRequest } from "../../hooks/useExpandNode";
+import { useUpdateNodeCountsQuery } from "../../hooks/useUpdateNodeCounts";
 
 export type NodeExpandContentProps = {
   classNamePrefix?: string;
   vertex: Vertex;
 };
 
-const NodeExpandContent = ({
+export default function NodeExpandContent({
   classNamePrefix = "ft",
   vertex,
-}: NodeExpandContentProps) => {
-  const t = useTranslations();
+}: NodeExpandContentProps) {
   const styleWithTheme = useWithTheme();
 
   return (
     <div className={styleWithTheme(defaultStyles(classNamePrefix))}>
       <VertexHeader vertex={vertex} classNamePrefix={classNamePrefix} />
-      {vertex.data.neighborsCount === 0 && (
-        <PanelEmptyState
-          icon={<GraphIcon />}
-          title={t("node-expand.no-connections-title")}
-          subtitle={t("node-expand.no-connections-subtitle")}
-        />
-      )}
-      {vertex.data.neighborsCount !== 0 && (
-        <NeighborDetails vertex={vertex} classNamePrefix={classNamePrefix} />
-      )}
+      <ExpandSidebarContent vertex={vertex} classNamePrefix={classNamePrefix} />
     </div>
   );
-};
+}
 
-function NeighborDetails({
+function ExpandSidebarContent({
+  classNamePrefix,
+  vertex,
+}: {
+  classNamePrefix?: string;
+  vertex: Vertex;
+}) {
+  const t = useTranslations();
+  const query = useUpdateNodeCountsQuery(vertex.data.id);
+
+  if (query.isError) {
+    return <PanelError error={query.error} onRetry={query.refetch} />;
+  }
+
+  if (query.isPending) {
+    return (
+      <PanelEmptyState icon={<LoadingSpinner />} title={"Loading Neighbors"} />
+    );
+  }
+
+  if (!query.data || query.data?.totalCount <= 0) {
+    return (
+      <PanelEmptyState
+        icon={<GraphIcon />}
+        title={t("node-expand.no-connections-title")}
+        subtitle={t("node-expand.no-connections-subtitle")}
+      />
+    );
+  }
+
+  return (
+    <>
+      <NeighborsList vertex={vertex} classNamePrefix={classNamePrefix} />
+      <ExpansionOptions classNamePrefix={classNamePrefix} vertex={vertex} />
+    </>
+  );
+}
+
+function ExpansionOptions({
   vertex,
   classNamePrefix,
 }: {
@@ -69,29 +98,29 @@ function NeighborDetails({
   const hasUnfetchedNeighbors = Boolean(vertex.data.__unfetchedNeighborCount);
   const hasSelectedType = Boolean(selectedType);
 
+  if (!hasUnfetchedNeighbors) {
+    return (
+      <PanelEmptyState
+        className={pfx("empty-panel-state")}
+        icon={<GraphIcon />}
+        title={t("node-expand.no-unfetched-title")}
+        subtitle={t("node-expand.no-unfetched-subtitle")}
+      />
+    );
+  }
+
   return (
     <>
-      <NeighborsList vertex={vertex} classNamePrefix={classNamePrefix} />
-      {!hasUnfetchedNeighbors && (
-        <PanelEmptyState
-          className={pfx("empty-panel-state")}
-          icon={<GraphIcon />}
-          title={t("node-expand.no-unfetched-title")}
-          subtitle={t("node-expand.no-unfetched-subtitle")}
-        />
-      )}
-      {hasUnfetchedNeighbors && (
-        <NodeExpandFilters
-          classNamePrefix={classNamePrefix}
-          neighborsOptions={neighborsOptions}
-          selectedType={selectedType}
-          onSelectedTypeChange={setSelectedType}
-          filters={filters}
-          onFiltersChange={setFilters}
-          limit={limit}
-          onLimitChange={setLimit}
-        />
-      )}
+      <NodeExpandFilters
+        classNamePrefix={classNamePrefix}
+        neighborsOptions={neighborsOptions}
+        selectedType={selectedType}
+        onSelectedTypeChange={setSelectedType}
+        filters={filters}
+        onFiltersChange={setFilters}
+        limit={limit}
+        onLimitChange={setLimit}
+      />
       <ModuleContainerFooter>
         <ExpandButton
           isDisabled={!hasUnfetchedNeighbors || !hasSelectedType}
@@ -140,5 +169,3 @@ function ExpandButton({
     </Button>
   );
 }
-
-export default NodeExpandContent;